// frontend/src/app/gen-api/schemas/postMessagingChatsBody.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Super Course SMS API
 * API documentation for the Super Course SMS services
 * OpenAPI spec version: 1.0.0
 */
import type { PostMessagingChatsBodyType } from './postMessagingChatsBodyType';

/**
 * Create a new chat
 */
export type PostMessagingChatsBody = {
  /** @minItems 2 */
  participants: string[];
  type?: PostMessagingChatsBodyType;
  name?: string;
};