// frontend/src/app/gen-api/schemas/postMessagingChatsChatIdReadBody.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Supercourse SMS API
 * API documentation for the Supercourse SMS service
 * OpenAPI spec version: 1.0.0
 */

/**
 * Mark all chat messages as read
 */
export type PostMessagingChatsChatIdReadBody = {
  /** @minLength 1 */
  chatId: string;
  /** @minLength 1 */
  userId: string;
};