// frontend/src/app/gen-api/schemas/postMessagingMessagesMessageIdDeliveredBody.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Supercourse SMS API
 * API documentation for the Supercourse SMS service
 * OpenAPI spec version: 1.0.0
 */

/**
 * Mark message as delivered
 */
export type PostMessagingMessagesMessageIdDeliveredBody = {
  /** @minLength 1 */
  messageId: string;
  /** @minLength 1 */
  userId: string;
};